+++ src/pages/config.vue
@@ -1,15 +1,14 @@
 <script setup lang="ts">
 import MultipleChoiceMenu from '../shared/components/multipleChoiceMenu.vue';
-import { MultipleChoiceMenuParams, ModalParams, UpdateDeleteCategory, GetUpdateProfile } from '../types';
+import { MultipleChoiceMenuParams, ModalParams, GetUpdateProfile } from '../types';
 import { ref } from 'vue';
 import { useDatabaseStore } from '../stores/databaseStore';
-import { useCategoryStore } from '../stores/categoryStore';
 import { useProfileStore } from '../stores/profileStore';
 import Modal from '../shared/components/modal.vue';
 import FormCreateCategory from '../shared/components/formCreateCategory.vue';
+import FormUpdateDeleteCategory from '../shared/components/formUpdateDeleteCategory.vue';
 
 const databaseStore = useDatabaseStore();
-const categoryStore = useCategoryStore();
 const profileStore = useProfileStore();
 
 const multipleChoiceMenuParams: MultipleChoiceMenuParams[] = [
@@ -20,20 +19,11 @@ const multipleChoiceMenuParams: MultipleChoiceMenuParams[] = [
 
 const categoryMenuChoice = ref<number>(1);
 
-const formDataUpdateDeleteCategory = ref<UpdateDeleteCategory>({
-  id: null,
-  category: '',
-  spend_target_value: 0,
-  initial_flag: true,
-});
-
 const modalParams = ref<ModalParams>({
   status: false,
   class: '',
   message: '',
 });
-
-const usedCategory = ref<boolean>(true);
 
 const updateProfile = ref<GetUpdateProfile>({
   target_value_total_price: profileStore.profile[0].target_value_total_price,
@@ -63,52 +53,10 @@ const callSetModalParams = ({ cssClass, message }: { cssClass: string; message: 
 
 const categoryMenuToggle = (index: number): void => {
   categoryMenuChoice.value = index;
-  formDataUpdateDeleteCategory.value = {
-    id: null,
-    category: '',
-    spend_target_value: 0,
-    initial_flag: true,
-  };
   updateProfile.value = {
     target_value_total_price: profileStore.profile[0].target_value_total_price,
     target_value_fixed_cost: profileStore.profile[0].target_value_fixed_cost,
     target_value_deferred_pay: profileStore.profile[0].target_value_deferred_pay,
-  };
-};
-
-const submitFormUpdateCategory = async (): Promise<void> => {
-  const value = formDataUpdateDeleteCategory.value;
-  try {
-    await databaseStore.updateCategory([value.category, value.spend_target_value, value.id!]);
-    await categoryStore.set(await databaseStore.getCategoryAll());
-    setModalParams({ cssClass: 'success', message: '項目の更新に成功しました。' });
-  } catch (error) {
-    console.error(error);
-    setModalParams({ cssClass: 'error', message: '項目の更新に失敗しました。' });
-  }
-  formDataUpdateDeleteCategory.value = {
-    id: null,
-    category: '',
-    spend_target_value: 0,
-    initial_flag: true,
-  };
-};
-
-const submitFormDeleteCategory = async (): Promise<void> => {
-  const value = formDataUpdateDeleteCategory.value;
-  try {
-    await databaseStore.deleteCategory(value.id!);
-    await categoryStore.set(await databaseStore.getCategoryAll());
-    setModalParams({ cssClass: 'success', message: '項目の削除に成功しました。' });
-  } catch (error) {
-    console.error(error);
-    setModalParams({ cssClass: 'error', message: '項目の削除に失敗しました。' });
-  }
-  formDataUpdateDeleteCategory.value = {
-    id: null,
-    category: '',
-    spend_target_value: 0,
-    initial_flag: true,
   };
 };
 
@@ -128,18 +76,6 @@ const submitFormUpdateProfile = async (): Promise<void> => {
     target_value_deferred_pay: profileStore.profile[0].target_value_deferred_pay,
   };
 };
-
-const formDisplaySetParams = async (): Promise<void> => {
-  const result = categoryStore.category.find((item) => item.id === formDataUpdateDeleteCategory.value.id)!;
-  const data = await databaseStore.usedCategory(result.id);
-  usedCategory.value = data[0].exists_flag === 1;
-  formDataUpdateDeleteCategory.value = {
-    id: result.id,
-    category: result.category,
-    spend_target_value: result.spend_target_value,
-    initial_flag: result.initial_flag,
-  };
-};
 </script>
 
 <template>
@@ -152,34 +88,7 @@ const formDisplaySetParams = async (): Promise<void> => {
     <FormCreateCategory @modal-params="callSetModalParams" />
   </div>
   <div v-if="categoryMenuChoice === 2">
-    <form>
-      <div class="form-row mb-20">
-        <div class="form-contents">
-          <label for="input-category">項目</label>
-          <select id="input-category" @change="formDisplaySetParams" v-model="formDataUpdateDeleteCategory.id" required>
-            <option v-for="category in categoryStore.category" :value="category.id" :key="category.id">
-              {{ category.category }}
-            </option>
-          </select>
-        </div>
-        <div v-if="formDataUpdateDeleteCategory.id" style="display: flex">
-          <div v-if="!formDataUpdateDeleteCategory.initial_flag && !usedCategory" class="form-contents">
-            <label for="input-category">項目</label>
-            <input type="text" v-model="formDataUpdateDeleteCategory.category" id="input-category" />
-          </div>
-          <div v-else class="form-contents">
-            <label for="input-category">項目</label>
-            <input type="text" v-model="formDataUpdateDeleteCategory.category" readonly class="read-only" id="input-category" />
-          </div>
-          <div class="form-contents">
-            <label for="input-spend-target-value">目標値</label>
-            <input type="number" v-model="formDataUpdateDeleteCategory.spend_target_value" id="input-spend-target-value" />
-          </div>
-          <button @click="submitFormUpdateCategory" type="button">更新</button>
-          <button v-if="!formDataUpdateDeleteCategory.initial_flag && !usedCategory" @click="submitFormDeleteCategory" type="button">削除</button>
-        </div>
-      </div>
-    </form>
+    <FormUpdateDeleteCategory @modal-params="callSetModalParams" />
   </div>
   <div v-if="categoryMenuChoice === 3">
     <form style="height: 180px">
